chore: Extract global options to doctest/parts/context/options.h
